% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcodes_scrapping.R
\name{all_pcodes_feature_servers}
\alias{all_pcodes_feature_servers}
\title{Scrap pcodes datasets from UNOCHA REST API}
\usage{
all_pcodes_feature_servers()
}
\value{
a dataframe containing all pcodes at the lowest administrative level available
}
\description{
Scrap pcodes datasets from UNOCHA REST API
}
\examples{
\dontrun{
all_pcodes_feature_servers()
}
}
