% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcodes_scrapping.R
\name{available_countries}
\alias{available_countries}
\title{List countries with pcodes available}
\usage{
available_countries()
}
\value{
data.frame - a dataframe containing with countries where pcodes are
    available.
}
\description{
List countries with pcodes available
}
\examples{
\dontrun{
available_countries()
}
}
