% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcodes_scrapping.R
\name{country_geojson_URL}
\alias{country_geojson_URL}
\title{Scrape geoJSON features on a specific country}
\usage{
country_geojson_URL(country_URL)
}
\arguments{
\item{country_URL}{character - A valid country FeatureServer directory URL.
Example: "https://gistmaps.itos.uga.edu/arcgis/rest/services/COD_External/ZMB_pcode/FeatureServer"}
}
\value{
geoJSON -  a geoJSON object containing the layer's features
}
\description{
Scrape geoJSON features on a specific country
}
\examples{
\dontrun{
country_geojson_URL("https://gistmaps.itos.uga.edu/arcgis/rest/services/COD_External/ZMB_pcode/FeatureServer")
}
}
