% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcodes_scrapping.R
\name{country_pcodes_URL}
\alias{country_pcodes_URL}
\title{Scrape all pcodes information on a specific country}
\usage{
country_pcodes_URL(country_URL)
}
\arguments{
\item{country_URL}{character - A valid country FeatureServer directory URL.
Example: "https://gistmaps.itos.uga.edu/arcgis/rest/services/COD_External/ZMB_pcode/FeatureServer"}
}
\value{
list - a list for the specified URL containing two elements:
    - name : character - string identifying the name of the layer
    - data : data.frame - data frame containing the layer data.
}
\description{
Scrape all pcodes information on a specific country
}
\examples{
\dontrun{
country_pcodes_URL("https://gistmaps.itos.uga.edu/arcgis/rest/services/COD_External/ZMB_pcode/FeatureServer")
}
}
