% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pcode_admin.R
\name{find_pcodes_admin}
\alias{find_pcodes_admin}
\title{Find pcode for a given administrative unit name}
\usage{
find_pcodes_admin(
  admin,
  country_iso3 = NULL,
  admin_level = c("admin1", "admin2", "admin3"),
  google_api_key = NULL,
  pcodes_df = NULL,
  language = NULL
)
}
\arguments{
\item{admin}{character or vector identifying administrative unit.}

\item{country_iso3}{character string with the ISO3 country code for the
country analysed. See [here](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3) for more details.}

\item{admin_level}{character string identifying a specific administrative
level where the pcode should be looked for. Three values are possible:
"admin1", "admin2", and "admin3". For instance, if you are looking to
pcode the region of Bamako in Mali, you should specify "admin1", as there
is also an administrative unit 2 with the same name. By default the
lowest administrative unit found will be returned}

\item{google_api_key}{a string with a google API key. See vignette("googleway-vignette") for more details.}

\item{pcodes_df}{dataframe containing matching key between pcodes and admin
names. For more details see [here](https://github.com/elliottmess/pcodesOCHA).}

\item{language}{character string identifying which language is being used.}
}
\value{
dataframe containing
}
\description{
Find pcode for a given administrative unit name
}
\examples{
find_pcodes_admin("Bamako", country_iso3 = "MLI", admin_level = "admin1")

}
