% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcodes_scrapping.R
\name{one_URL_geoJSON}
\alias{one_URL_geoJSON}
\title{Scrap geoJSON from one URL on \href{https://gistmaps.itos.uga.edu/arcgis/rest/services/COD_External}{OCHA COD REST API}}
\usage{
one_URL_geoJSON(layer_URL)
}
\arguments{
\item{layer_URL}{character - A valid FeatureServer layer directory URL.
Example: "https://gistmaps.itos.uga.edu/arcgis/rest/services/COD_External/ZMB_pcode/FeatureServer/2"}
}
\value{
geoJSON -  a geoJSON object containing the layer's features
}
\description{
Scrap geoJSON from one URL on \href{https://gistmaps.itos.uga.edu/arcgis/rest/services/COD_External}{OCHA COD REST API}
}
\examples{
\dontrun{
one_URL_geoJSON("https://gistmaps.itos.uga.edu/arcgis/rest/services/COD_External/ZMB_pcode/FeatureServer/2")
}
}
