% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcodes_scrapping.R
\name{one_URL_properties}
\alias{one_URL_properties}
\title{Scrap data properties from one URL on OCHA COD API}
\usage{
one_URL_properties(layer_URL)
}
\arguments{
\item{layer_URL}{character - A valid FeatureServer layer directory URL.
Example: "https://gistmaps.itos.uga.edu/arcgis/rest/services/COD_External/ZMB_pcode/FeatureServer/2"}
}
\value{
list - a list containing two elements:
    - name : character - string identifying the name of the layer
    - data : data.frame - data frame containing the layer data.
}
\description{
Scrap data properties from one URL on OCHA COD API
}
\examples{
\dontrun{
one_URL_properties("https://gistmaps.itos.uga.edu/arcgis/rest/services/COD_External/ZMB_pcode/FeatureServer/2")
}
}
