% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcodes_scrapping.R
\name{rename_pcodes}
\alias{rename_pcodes}
\title{Rename pcode dataframe with uniform column names}
\usage{
rename_pcodes(pcodes)
}
\arguments{
\item{pcodes}{data.frame containing pcodes extracted from OCHA server}
}
\value{
dataframe with uniform column names
}
\description{
Rename pcode dataframe with uniform column names
}
\examples{

}
